+++ src/main/java/roomescape/controller/api/UserReservationTimeController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import roomescape.controller.dto.FindTimeAndAvailabilityResponse;
+import roomescape.controller.dto.response.FindTimeAndAvailabilityResponse;
 import roomescape.service.ReservationTimeService;
 
 @RestController
