+++ src/main/java/roomescape/controller/api/UserThemeController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import roomescape.controller.dto.FindThemeResponse;
+import roomescape.controller.dto.response.FindThemeResponse;
 import roomescape.service.ThemeService;
 
 @RestController
