+++ src/main/java/roomescape/controller/dto/response/ErrorMessageResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.response;
 
 public record ErrorMessageResponse(String message) {
 }
