+++ src/main/java/roomescape/controller/dto/response/FindMemberResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.response;
 
 import roomescape.domain.member.Member;
