+++ src/main/java/roomescape/controller/dto/response/FindReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.response;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
