+++ src/main/java/roomescape/controller/dto/response/FindTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.response;
 
 import java.time.LocalTime;
 
@@ -7,7 +7,6 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import roomescape.domain.reservation.ReservationTime;
 
 public record FindTimeResponse(Long id, @JsonFormat(pattern = "HH:mm") LocalTime startAt) {
-
     public static FindTimeResponse from(ReservationTime time) {
         return new FindTimeResponse(
                 time.getId(),
