+++ src/main/java/roomescape/controller/dto/response/PaymentErrorMessageResponse.java
@@ -1,5 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.response;
 
 public record PaymentErrorMessageResponse(String code, String message) {
-
 }
