+++ src/test/java/roomescape/service/MemberServiceTest.java
@@ -16,7 +16,7 @@ import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 
 import io.restassured.RestAssured;
-import roomescape.controller.dto.FindMemberResponse;
+import roomescape.controller.dto.response.FindMemberResponse;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.Role;
 import roomescape.global.exception.RoomescapeException;
