+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -23,9 +23,9 @@ import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 
 import io.restassured.RestAssured;
-import roomescape.controller.dto.CreateTimeResponse;
-import roomescape.controller.dto.FindTimeAndAvailabilityResponse;
-import roomescape.controller.dto.FindTimeResponse;
+import roomescape.controller.dto.response.CreateTimeResponse;
+import roomescape.controller.dto.response.FindTimeAndAvailabilityResponse;
+import roomescape.controller.dto.response.FindTimeResponse;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.Role;
 import roomescape.domain.reservation.Reservation;
